requirements.txt
numpy
Pillow
matplotlib
scipy
scikit-image
scikit-learn
pandas
seaborn
PyYAML
torch
torchvision
lpips
